+++ server/src/group/group.controller.ts
@@ -1,8 +1,9 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Session } from '@nestjs/common';
 import { GroupService } from './group.service';
 import { CreateGroupDto } from './dto/create-group.dto';
 import { UpdateGroupDto } from './dto/update-group.dto';
 import { ApiOkResponse, ApiTags } from '@nestjs/swagger';
+import session from 'express-session';
 
 @Controller('group')
 @ApiTags("体检套餐接口")
@@ -10,8 +11,11 @@ export class GroupController {
   constructor(private readonly groupService: GroupService) {}
 
   @Post()
-  @ApiOkResponse({ description: '创建体检套餐' })
-  create(@Body() createGroupDto: CreateGroupDto) {
+  @ApiOkResponse({ description: '创建体检套餐，需要管理员权限' })
+  create(@Body() createGroupDto: CreateGroupDto, @Session() session) {
+    if(session.loginType !== "admin") {
+      throw new Error("权限不足");
+    }
     return this.groupService.create(createGroupDto);
   }
 
@@ -31,14 +35,20 @@ export class GroupController {
   }
 
   @Patch(':id')
-  @ApiOkResponse({ description: '修改单个体检套餐' })
-  update(@Param('id') id: number, @Body() updateGroupDto: UpdateGroupDto) {
+  @ApiOkResponse({ description: '修改单个体检套餐，需要管理员权限' })
+  update(@Param('id') id: number, @Body() updateGroupDto: UpdateGroupDto, @Session() session) {
+    if(session.loginType !== "admin") {
+      throw new Error("权限不足");
+    }
     return this.groupService.update(+id, updateGroupDto);
   }
 
   @Delete(':id')
-  @ApiOkResponse({ description: '删除单个体检套餐' })
-  remove(@Param('id') id: string) {
+  @ApiOkResponse({ description: '删除单个体检套餐，需要管理员权限' })
+  remove(@Param('id') id: string, @Session() session) {
+    if(session.loginType !== "admin") {
+      throw new Error("权限不足");
+    }
     return this.groupService.remove(+id);
   }
 }
